// src/data/dto.ts
export interface DataCreateDTO {
  readonly id: number
  readonly code: number
  readonly type: string
  readonly name: string
  readonly rayon: string
  readonly adress: string
  readonly contact: string
  readonly k1: string
  readonly k2: string
  readonly link: string
  readonly gorpit: string
  readonly director: string
  readonly programm: string
  readonly deyatelnost: string
  readonly scoc: string
  readonly trosta: string
  readonly itcube: string
  readonly kvant: string
  readonly masterskay: string
}